Refactored favorite opponent and rival
